requirements.txt
fastapi
uvicorn
python-doctr
torch
torchvision
pillow
python-multipart